Added or modified .c files
I divided the project in 4 parts: file utilities, UI elements
(text-based), linked list utilities and the main.
The linked list utilities are functions to operate on a linked list,
while the UI elements are functions that take information from the user
and translate them in a way that they can be used in a linked list.
Basically, think of the UI elements as the middle man between the main
and the linked list.